orchestrator/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
python-dotenv==1.0.1
aiosqlite==0.20.0
openai==1.58.1
google-generativeai==0.8.4
numpy>=2.0.0
opencv-python-headless==4.10.0.84
hsemotion-onnx>=0.3.0
torch>=2.6.0
transformers>=4.47.0
Pillow>=11.0.0
